+++ xtask/src/commands/mod.rs
@@ -2,9 +2,11 @@ pub(crate) use dist::Dist;
 pub(crate) use package::Package;
 pub(crate) use prep::Prep;
 pub(crate) use publish_npm::PublishNpm;
+pub(crate) use register_oauth_client::RegisterOauthClient;
 
 pub(crate) mod dist;
 pub(crate) mod package;
 pub(crate) mod prep;
 pub(crate) mod publish_npm;
+pub(crate) mod register_oauth_client;
 pub(crate) mod version;